Remove letter-spacing
The letter-spacing, especially on headings, is very distracting. It makes the website look less professional. If more emphasis is needed, then either a more distinct font should be used or some other graphical treatment (underlines, colors, etc.) might be more appropriate. But I think it the headings are distinct enough with just the bold & size treatment already present in the CSS.